googlemap
Former-commit-id: 99903adc009c8aaa415afc7fd38a04e0c78e2ce7
Former-commit-id: 26a3e36962d53c9d19885735dcd97b1daae40cdc [formerly 4f56e50347959f56f54a4c196e068d3339a3b288]
Former-commit-id: 8bd5ddc5fc202e86c78bc102f84c92be28e6f1bb
Former-commit-id: d5ed9bd64258893fbc64eb03907e82a740149f4b

+++ googlemap/ios/Classes/AkylasGMSURLTileLayer.m
@@ -240,7 +240,7 @@ uint64_t TileKey(NSUInteger theX, NSUInteger theY, NSUInteger theZ)
         [receiver receiveTileWithX:x y:y zoom:zoom image:nil];
         return;
     }
-    if (_minZoom >= 0 && zoom < _minZoom) {
+    if (_minZoom >= 0 && (CGFloat)zoom  < _minZoom) {
         [receiver receiveTileWithX:x y:y zoom:zoom image:self.errorImage];
         return;
     }
